backend/requirements.txt
Flask==3.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
MarkupSafe==2.1.3
PyJWT==2.8.0
SQLAlchemy==2.0.25
Werkzeug==3.0.1 # or 2.0.3
python-dotenv==1.0.1
pytest==8.0.0
# Flask-3.0.1 Flask-Cors-4.0.0 Flask-JWT-Extended-4.5.3 Flask-Login-0.6.3 Flask-SQLAlchemy-3.1.1 Jinja2-3.1.3 MarkupSafe-2.1.3 PyJWT-2.8.0 SQLAlchemy-2.0.25 Werkzeug-3.0.1 blinker-1.7.0 click-8.1.7 iniconfig-2.0.0 itsdangerous-2.1.2 packaging-24.0 pluggy-1.4.0 pytest-8.0.0 python-dotenv-1.0.1 typing-extensions-4.10.0